move around some functions to utils
